+++ cmd/agent/entrypoint.go
@@ -17,6 +17,7 @@ import (
 	"github.com/grafana/agent/pkg/metrics/instance"
 	"github.com/grafana/agent/pkg/server"
 	"github.com/grafana/agent/pkg/traces"
+	"github.com/grafana/agent/pkg/usagestats"
 	"github.com/oklog/run"
 	"google.golang.org/grpc"
 	"gopkg.in/yaml.v2"
@@ -42,6 +43,7 @@ type Entrypoint struct {
 	lokiLogs     *logs.Logs
 	tempoTraces  *traces.Traces
 	integrations config.Integrations
+	reporter     *usagestats.Reporter
 
 	reloadListener net.Listener
 	reloadServer   *http.Server
@@ -84,7 +86,13 @@ func NewEntrypoint(logger *server.Logger, cfg *config.Config, reloader Reloader)
 	if err != nil {
 		return nil, err
 	}
+
 	ep.integrations, err = config.NewIntegrations(logger, &cfg.Integrations, integrationGlobals)
+	if err != nil {
+		return nil, err
+	}
+
+	ep.reporter, err = usagestats.NewReporter(logger, cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -310,6 +318,17 @@ func (ep *Entrypoint) Start() error {
 	}, func(e error) {
 		srvCancel()
 	})
+
+	ep.mut.Lock()
+	cfg := ep.cfg
+	ep.mut.Unlock()
+	if cfg.EnableUsageReport {
+		g.Add(func() error {
+			return ep.reporter.Start(srvContext)
+		}, func(e error) {
+			srvCancel()
+		})
+	}
 
 	go func() {
 		for range notifier {